+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG - Link Duplicates
 
+## v0.1.6 - improved handling of files to ignore
+
+- ignore symlinks
+- ignore files of zero size completely
+
 ## v0.1.5
 
 - avoid crashing out with a `StopIteration` when path contains multiple files with zero size